Rewind: Update rewind state when changing credentials
The credentials banner and credentials form are still dependent on old
API calls we have decided to stop relying on. Their logic is replicated
in here in an almost-identical way as it is in the server, but it's not
entirely the same, plus it's duplicated.

In D9399 we're returning the `rewind_state` as an additional property on
calls to the credentials endpoints so that we can start gradually
replacing the old API calls with the centralized `/rewind` API.

This patch incorporates our ability to read and inject that response
into Calypso. It shouldn't have any effect until the server starts
sending those updates.